rename icons folder
