+++ src/Components/About/AboutControl.jsx
@@ -6,7 +6,7 @@ import Dialog from '../Dialog'
 import {ControlButton} from '../Buttons'
 import AboutGuide from './AboutGuide'
 import PrivacyControl from './PrivacyControl'
-import AboutIcon from '../../assets/2D_Icons/Information.svg'
+import AboutIcon from '../../assets/icons/Information.svg'
 import LogoB from '../../assets/LogoB.svg'
