+++ src/Components/About/AboutGuide.jsx
@@ -2,10 +2,10 @@ import React, {useContext} from 'react'
 import Box from '@mui/material/Box'
 import {useTheme} from '@mui/styles'
 import {ColorModeContext} from '../../Context/ColorMode'
-import CutPlaneIcon from '../../assets/2D_Icons/CutPlane.svg'
-import ListIcon from '../../assets/2D_Icons/List.svg'
-import NotesIcon from '../../assets/2D_Icons/Notes.svg'
-import ShareIcon from '../../assets/2D_Icons/Share.svg'
+import CutPlaneIcon from '../../assets/icons/CutPlane.svg'
+import ListIcon from '../../assets/icons/List.svg'
+import NotesIcon from '../../assets/icons/Notes.svg'
+import ShareIcon from '../../assets/icons/Share.svg'
 import LogoBuildings from '../../assets/Logo_Buildings.svg'
