+++ src/Components/ItemProperties/ItemPropertiesDrawer.jsx
@@ -5,7 +5,7 @@ import Typography from '@mui/material/Typography'
 import {TooltipIconButton} from '../Buttons'
 import {MOBILE_WIDTH} from '../Hooks'
 import {preprocessMediaQuery} from '../../utils/mediaQuery'
-import CloseIcon from '../../assets/2D_Icons/Close.svg'
+import CloseIcon from '../../assets/icons/Close.svg'
 
 
 /**
