+++ src/Components/MenuButton.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Box from '@mui/material/Box'
 import IconButton from '@mui/material/IconButton'
 import Tooltip from '@mui/material/Tooltip'
-import Hamburger from '../assets/2D_Icons/menu.svg'
+import Hamburger from '../assets/icons/menu.svg'
 
 
 const MenuButton = ({onClick}) => {
