+++ src/Components/NoContent.jsx
@@ -2,7 +2,7 @@ import React, {useContext} from 'react'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import {ColorModeContext} from '../Context/ColorMode'
-import AttentionIcon from '../assets/2D_Icons/Attention.svg'
+import AttentionIcon from '../assets/icons/Attention.svg'
 
 
 /**
