+++ src/Components/OperationsGroup.jsx
@@ -9,11 +9,11 @@ import CameraControl from './CameraControl'
 import CutPlaneMenu from './CutPlaneMenu'
 import ShareControl from './ShareControl'
 import {TooltipIconButton} from './Buttons'
-import ClearIcon from '../assets/2D_Icons/Clear.svg'
-import ListIcon from '../assets/2D_Icons/List.svg'
-import MoonIcon from '../assets/2D_Icons/Moon.svg'
-import NotesIcon from '../assets/2D_Icons/Notes.svg'
-import SunIcon from '../assets/2D_Icons/Sun.svg'
+import ClearIcon from '../assets/icons/Clear.svg'
+import ListIcon from '../assets/icons/List.svg'
+import MoonIcon from '../assets/icons/Moon.svg'
+import NotesIcon from '../assets/icons/Notes.svg'
+import SunIcon from '../assets/icons/Sun.svg'
 
 
 /**
